Freeze package versions (Straight) for full reproducibility
Formerly, this Emacs config could get out of sync with publicly
visible repos used in the `public` branch that are different from the
local repos used in the `main` branch, because the config is only
guaranteed to be compatible with those local checked out repos, and
upstream changes not reflected locally could therefore cause the
config to break.

This fixes the issue by using Straight's ability to "freeze" versions,
so that the config working at the time of the commit is proof that it
will always work (assuming those versions are still accessible in the
future of course). In order to keep the config both consistent as well
as up to date, the `public` branch will need to be periodically
updated and the versions refrozen once any issues are resolved.