make racket leader accessible in symex mode
As Lithium passes through undefined keys, it falls to interpretation
in vanilla emacs. But leader is fundamentally an Evil concept, so as
an alternative to declaring it as a binding in the user symex lithium
mode overrides, it seems OK to couple it to the evil state we've
already defined.