move symex-rigpa config to come after rigpa
It appears that the use-package `:after` keyword only controls the
order of *loading* and the evaluation of configuration. If Straight
encounters symex-rigpa before it has seen rigpa, it will try to find
the latter on package archives and then complain that the dependency
cannot be found.

This overloading of installation and configuration using the same
macro is perhaps not a good idea. It might also be valuable to be able
to declare a local recipe repository and add that to straight, so that
the order of appearance in config would not matter for installation,
as the local repository would be consulted in each case.