symex: remove some UI niceties
These modes are *global* modes, even though their operation is within
specific buffers. Unfortunately, this means that we cannot reliably
toggle them on and off upon entry and exit from Symex mode, as we
might be in Symex mode in one buffer but not another buffer, and the
UI features being toggled globally cannot support the expected
behavior in such cases without the introduction of additional hooks.